Render red sphere & apect ratio fix

+++ src/main.rs
@@ -8,7 +8,7 @@ fn main() {
     // Set up the image parameters
     let aspect_ratio = 16.0 / 9.0;
     let img_width: u32 = 400;
-    let img_height = img_width / aspect_ratio as u32;
+    let img_height = (img_width as f64 / aspect_ratio) as u32;
 
     // Set up the view parameters
     let view_height = 2.0;
